v1.05
- StatusEffect class added to track buffs;
- combos added;
- UI for combos done;
- reflective shield reflects alien projectiles;

+++ src/entities/aliens/AlienShotEntity.java
@@ -2,6 +2,7 @@ package entities.aliens;
 
 import java.awt.Graphics2D;
 
+import utility.image.AnimationManager;
 import utility.image.ImageManager;
 import entities.Entity;
 import entities.ShipEntity;
@@ -10,6 +11,8 @@ import main.Game;
 public class AlienShotEntity extends Entity{
 
 	private int type;
+	// reflected by the player shield; reflected bullets cannot be reflected again
+	private boolean reflected = false;
 	
 	public AlienShotEntity(Game game, int x, int y, int type){
 		super(x, y);
@@ -29,6 +32,15 @@ public class AlienShotEntity extends Entity{
 	
 	public void collidedWith(Entity other){
 		if( other instanceof ShipEntity){
+			if (((ShipEntity) other).isInvulnerable()){
+				if (!reflected) {
+					dx = (x - ((ShipEntity) other).getCenteredX())*10;
+					dy = -dy;
+					reflected = true;
+					AnimationManager.getAnimationManager().spawnAnimation("effects/reflectionSparks", (int)x-10, (int)y-15, 2.0);
+				}
+				return;
+			}
 			game.getEntityManager().removeEntity(this);
 			game.notifyDeath();
 		}
@@ -36,7 +48,11 @@ public class AlienShotEntity extends Entity{
 
 	@Override
 	public void draw(Graphics2D g) {
-		g.drawImage(ImageManager.getImage("projectiles/alienShot" + type), (int)x, (int)y, null);
+		int width = ImageManager.getImage("projectiles/alienShot" + type).getWidth(null);
+		int height = ImageManager.getImage("projectiles/alienShot" + type).getHeight(null);
+		// reflected bullets are facing up i.e. image is flipped using negative height
+		height = reflected ? -height : height;
+		g.drawImage(ImageManager.getImage("projectiles/alienShot" + type), (int)x, (int)y, width, height, null);
 		
 	}
